requirements.txt
discord.py
sqlalchemy
pandas
python-dotenv
scapy
psycopg2-binary